network_of_assistants/noa-file-assistant/pyproject.toml
[project]
name = "noa-file-assistant"
version = "0.1.0"
description = "An assistant agent specialized in retrieving information from files."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "slim",
    "llama-index>=0.12.29",
    "llama-index-llms-azure-openai>=0.3.2",
    "llama-index-llms-ollama>=0.5.4",
    "llama-index-llms-openai-like>=0.3.4",
    "ioa-observe-sdk>=1.0.9",
    "slim-bindings>=0.3.5",
]

[tool.uv.sources]
slim = { workspace = true }

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project.scripts]
noa-file-assistant = "main:run"
